chore: refactor all queries into repos
