+++ app/projects/[id]/page.tsx
@@ -1,19 +1,19 @@
 'use server';
 import Tasks from '@/app/tasks/_components/tasks';
-import { getProjects, getTasks } from '@/db/queries';
+import { projectRepo } from '@/db/repos/project-repo';
+import { taskRepo } from '@/db/repos/task-repo';
 
 const workspaceId = 'work';
 export default async function Page(props: { params: Promise<{ id: string }> }) {
   const params = await props.params;
-  // TODO: fix overfetching
-  const projects = await getProjects({ workspaceId });
+  const projects = await projectRepo.getMany({ workspaceId });
 
   const project = projects.find((p) => p.id === params.id);
   if (!project) {
     return <div>Project not found</div>;
   }
 
-  const tasks = await getTasks({ workspaceId, projectId: params.id });
+  const tasks = await taskRepo.getMany({ workspaceId, projectId: params.id });
 
   return (
     <div>
