+++ app/projects/actions.ts
@@ -1,13 +1,10 @@
 'use server';
-import { db } from '@/db';
-import { projects } from '@/db/schema/projects';
-import { eq } from 'drizzle-orm';
+import { projectRepo } from '@/db/repos/project-repo';
 import { revalidatePath } from 'next/cache';
 
 export const deleteProject = async (id: string) => {
-  // TODO: authorization
   try {
-    await db.delete(projects).where(eq(projects.id, id));
+    await projectRepo.del(id);
     revalidatePath('/');
   } catch (e) {
     console.error(e);
