+++ app/tasks/actions.ts
@@ -1,13 +1,10 @@
 'use server';
-import { db } from '@/db';
-import { tasks } from '@/db/schema/tasks';
-import { eq } from 'drizzle-orm';
+import { taskRepo } from '@/db/repos/task-repo';
 import { revalidatePath } from 'next/cache';
 
 export const deleteTask = async (id: string) => {
-  // TODO: authorization
   try {
-    await db.delete(tasks).where(eq(tasks.id, id));
+    await taskRepo.del(id);
     revalidatePath('/');
   } catch (e) {
     console.log(e);
